Set default widget; remove secondary widgets from manifest, update configs
