+++ components/widgets/fires/burned-area/index.js
@@ -2,8 +2,6 @@ import { all, spread } from 'axios';
 import uniq from 'lodash/uniq';
 import moment from 'moment';
 
-import firesAlertsStats from 'components/widgets/fires/fires-alerts';
-
 import { fetchBurnedArea, fetchVIIRSLatest } from 'services/analysis-cached';
 
 import {
@@ -18,8 +16,7 @@ import {
 
 import getWidgetProps from './selectors';
 
-const defaultConfig = {
-  ...firesAlertsStats,
+export default {
   widget: 'burnedAreaStats',
   title: 'Weekly Burned Area in {location}',
   large: true,
@@ -55,6 +52,9 @@ const defaultConfig = {
     },
   ],
   refetchKeys: ['dataset', 'forestType', 'landCategory'],
+  preventRenderKeys: ['startIndex', 'endIndex'],
+  admins: ['adm0', 'adm1', 'adm2'],
+  chartType: 'composedChart',
   datasets: [
     {
       dataset: POLITICAL_BOUNDARIES_DATASET,
@@ -67,9 +67,14 @@ const defaultConfig = {
       layers: [BURNED_AREA_MODIS],
     },
   ],
+  hideLayers: true,
   visible: ['dashboard', 'analysis'],
   types: ['country'],
-  metaKey: 'widget_fire_alert_location',
+  metaKey: '',
+  sortOrder: {
+    summary: 100,
+    fires: 1,
+  },
   settings: {
     dataset: 'modis_burned_area',
   },
@@ -81,6 +86,215 @@ const defaultConfig = {
       'Fires burned {area} of land between {start_date} and {end_date}, when data were most recently available. The area burned during this time period is {status} compared to the area burned in previous years going back to {dataset_start_year}.',
     allBurnWithInd:
       'Fires burned {area} of land within {indicator} between {start_date} and {end_date}, when data were most recently available. The area burned during this time period is {status} compared to the area burned in previous years going back to {dataset_start_year}.',
+  },
+  whitelists: {
+    adm0: [
+      'AFG',
+      'AGO',
+      'ALB',
+      'AND',
+      'ANT',
+      'ARE',
+      'ARG',
+      'ARM',
+      'AUS',
+      'AUT',
+      'AZE',
+      'BDI',
+      'BEL',
+      'BEN',
+      'BFA',
+      'BGD',
+      'BGR',
+      'BHR',
+      'BHS',
+      'BIH',
+      'BLM',
+      'BLR',
+      'BLZ',
+      'BOL',
+      'BRA',
+      'BRB',
+      'BRN',
+      'BTN',
+      'BWA',
+      'CAF',
+      'CAN',
+      'CHE',
+      'CHL',
+      'CHN',
+      'CIV',
+      'CMR',
+      'COD',
+      'COG',
+      'COL',
+      'COM',
+      'CPV',
+      'CRI',
+      'CUB',
+      'CYP',
+      'CZE',
+      'DEU',
+      'DJI',
+      'DMA',
+      'DNK',
+      'DOM',
+      'DZA',
+      'ECU',
+      'EGY',
+      'ERI',
+      'ESP',
+      'EST',
+      'ETH',
+      'FIN',
+      'FJI',
+      'FLK',
+      'FRA',
+      'FSM',
+      'GAB',
+      'GBR',
+      'GEO',
+      'GHA',
+      'GIB',
+      'GIN',
+      'GLP',
+      'GMB',
+      'GNB',
+      'GNQ',
+      'GRC',
+      'GRL',
+      'GTM',
+      'GUF',
+      'GUM',
+      'GUY',
+      'HND',
+      'HRV',
+      'HTI',
+      'HUN',
+      'IDN',
+      'IND',
+      'IRL',
+      'IRN',
+      'IRQ',
+      'ISR',
+      'ITA',
+      'JAM',
+      'JOR',
+      'JPN',
+      'KAZ',
+      'KEN',
+      'KGZ',
+      'KHM',
+      'KIR',
+      'KNA',
+      'KOR',
+      'KWT',
+      'LAO',
+      'LBN',
+      'LBR',
+      'LBY',
+      'LCA',
+      'LIE',
+      'LKA',
+      'LSO',
+      'LTU',
+      'LUX',
+      'LVA',
+      'MAR',
+      'MCO',
+      'MDA',
+      'MDG',
+      'MDV',
+      'MEX',
+      'MHL',
+      'MKD',
+      'MLI',
+      'MLT',
+      'MMR',
+      'MNE',
+      'MNG',
+      'MNP',
+      'MOZ',
+      'MRT',
+      'MSR',
+      'MTQ',
+      'MUS',
+      'MWI',
+      'MYS',
+      'NAM',
+      'NCL',
+      'NER',
+      'NGA',
+      'NIC',
+      'NLD',
+      'NOR',
+      'NPL',
+      'NZL',
+      'OMN',
+      'PAK',
+      'PAN',
+      'PCN',
+      'PER',
+      'PHL',
+      'PNG',
+      'POL',
+      'PRI',
+      'PRK',
+      'PRT',
+      'PRY',
+      'PSE',
+      'PYF',
+      'QAT',
+      'REU',
+      'ROU',
+      'RUS',
+      'RWA',
+      'SAU',
+      'SDN',
+      'SEN',
+      'SGP',
+      'SLB',
+      'SLE',
+      'SLV',
+      'SOM',
+      'SRB',
+      'SSD',
+      'STP',
+      'SUR',
+      'SVK',
+      'SVN',
+      'SWE',
+      'SWZ',
+      'SYR',
+      'TCD',
+      'TGO',
+      'THA',
+      'TJK',
+      'TKL',
+      'TKM',
+      'TLS',
+      'TON',
+      'TTO',
+      'TUN',
+      'TUR',
+      'TUV',
+      'TZA',
+      'UGA',
+      'UKR',
+      'URY',
+      'USA',
+      'UZB',
+      'VAT',
+      'VEN',
+      'VIR',
+      'VNM',
+      'VUT',
+      'WSM',
+      'YEM',
+      'ZAF',
+      'ZMB',
+      'ZWE',
+    ],
   },
   getData: (params) =>
     all([fetchBurnedArea(params), fetchVIIRSLatest(params)]).then(
@@ -115,15 +329,3 @@ const defaultConfig = {
   getDataURL: (params) => [fetchBurnedArea({ ...params, download: true })],
   getWidgetProps,
 };
-
-export default {
-  widget: 'burnedAreaStats',
-  proxy: true,
-  refetchKeys: ['dataset'],
-  getWidget: (props) => { // called when settings changes
-    if (!props || !props.dataset) return defaultConfig;
-    if (props.dataset === 'modis_burned_area') return defaultConfig;
-    if (props.dataset === 'modis') return firesAlertsStats;
-    return defaultConfig;
-  },
-}
